Removing 1$ loans (#40)
* Updated Guide Page

* Updated Profile and Verification

Device Verified users are now unverified

* Updated Loan conditions

Only ORB verified users can get a loan from now on.

* Implement the plan

This commit implements the changes as per the provided plan.

* Merged Verification into Profile

* Loan Functionality

- ORB verified users now see only 10$ loans
- Device verified users see new error

---------
