+++ src/components/Header.tsx
@@ -113,12 +113,6 @@ export const Header = ({ title, showBack = true }: HeaderProps) => {
               </DropdownMenuItem>
               <DropdownMenuItem
                 className="min-h-[40px] focus:text-white focus-visible:text-white"
-                onClick={() => navigate("/upgrade-verification")}
-              >
-                Verification Level
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                className="min-h-[40px] focus:text-white focus-visible:text-white"
                 onClick={() => navigate("/loan-history")}
               >
                 Loan History
